Fixed bug in conversion.d with UTF sequences beginning with ?=.
In cases where the UTF sequence began with a "?=", this was taken as the end of the squence, and the contents between the start and the end, which was zero bytes long was sent to be decoded.  decode function amended to search for the ending '?=' one space AFTER the beginning of the sequence to stop confusing ?= which begins a UTF character with ?= which ends a UTF sequence.